+++ src/components/layouts/StandardLayout.js
@@ -3,7 +3,7 @@ import { Outlet } from 'react-router-dom';
 import Header from '../header/Header';
 import WithModals from '../common/modals/WithModals';
 import { ModalContextProvider } from '../../contexts/modal.context';
-import Toasts from '../common/Toasts';
+import Toasts from '../common/toasts/Toasts';
 
 const StandardLayout = () => {
   return (
