+++ src/components/pages/LandingPage.js
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { Navigate, useNavigate } from 'react-router-dom';
 
 import Loading from '../common/Loading';
-import Button from '../common/Button';
+import Button from '../common/buttons/Button';
 
 import { useAuthState } from '../../hooks/auth.hooks';
